+++ src/durabletask/azext_durabletask/commands.py
@@ -11,3 +11,4 @@ def load_command_table(self, _):  # pylint: disable=unused-argument
         self.command_table["durabletask retention-policy create"] = CreatePolicy(loader=self)
     with self.command_group('durabletask orchestration') as g:
         g.custom_command('list', 'list_orchestrations')
+        g.custom_command('show', 'show_orchestration')