+++ src/durabletask/HISTORY.rst
@@ -4,7 +4,7 @@ Release History
 ===============
 1.0.0b4
 +++++
-* Adding orchestration command to list orchestrations.
+* Adding orchestration command to list and show orchestrations.
 
 1.0.0b4
 +++++